dogs/requirements.txt
fastapi==0.109.2
uvicorn==0.27.0.post1
anthropic==0.44.0
python-dotenv==1.0.1
pydantic==2.6.1
requests==2.31.0
python-multipart==0.0.9
typing-extensions
starlette==0.36.3
pymysql